Model: split cached event types from display summaries
Structure-only commit: model/events.rs holds what collectors emit
(cache-relevant via FileEvents, noted in its doc), model/summary.rs
holds what the analyzer produces for display, and model.rs keeps the
shared core (Provider, SourceKind, TokenUsage) plus glob re-exports —
every crate::model::* path is unchanged. A cache-layout change and a
display-shape change now name different files in a diff.
